Renamed the RegistryLCManager to LCManager
